# api/NeoWallet_OpenAPI_v1.yaml
openapi: 3.1.0
info:
  title: NeoWallet API
  version: 1.0.0
  description: NeoWallet MVP API Contract Specification
  contact:
    name: NeoWallet API Team
    email: api@neowallet.com
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.neowallet.com/api/v1
    description: Production
  - url: https://api-staging.neowallet.com/api/v1
    description: Staging
  - url: https://api-dev.neowallet.com/api/v1
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - requestId
        - timestamp
      properties:
        code:
          type: string
          example: BUDGET_NOT_FOUND
        message:
          type: string
          example: Budget was not found.
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    User:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        phoneNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        familyId:
          type: string
          format: uuid
          nullable: true
        familyRole:
          type: string
          enum: [OWNER, MEMBER, RESTRICTED]
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
    DeleteUserRequest:
      type: object
      required:
        - confirmation
      properties:
        confirmation:
          type: string
    UserPreferences:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        locale:
          type: string
        currency:
          type: string
        timezone:
          type: string
        notificationPreferences:
          type: object
          properties:
            budgetAlerts:
              type: boolean
            billReminders:
              type: boolean
            savingsUpdates:
              type: boolean
            financialHealthUpdates:
              type: boolean
        aiPreferences:
          type: object
          properties:
            responseStyle:
              type: string
              enum: [CONCISE, DETAILED]
            language:
              type: string
    UpdateUserPreferencesRequest:
      type: object
      properties:
        locale:
          type: string
        currency:
          type: string
        timezone:
          type: string
        notificationPreferences:
          type: object
          properties:
            budgetAlerts:
              type: boolean
            billReminders:
              type: boolean
            savingsUpdates:
              type: boolean
            financialHealthUpdates:
              type: boolean
        aiPreferences:
          type: object
          properties:
            responseStyle:
              type: string
              enum: [CONCISE, DETAILED]
            language:
              type: string
...
paths:
...
  /users/me:
    get:
      summary: Get current user profile
      operationId: getCurrentUser
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update current user profile
      operationId: updateCurrentUser
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Phone number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete current user account
      operationId: deleteCurrentUser
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteUserRequest'
      responses:
        '204':
          description: User account deleted successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cannot delete family owner with members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/me/preferences:
    get:
      summary: Get user preferences
      operationId: getUserPreferences
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update user preferences
      operationId: updateUserPreferences
      tags:
        - Users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserPreferencesRequest'
      responses:
        '200':
          description: User preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...